# openapi.yaml
openapi: 3.1.0
info:
  title: ZenStack Generated API
  version: 1.0.0
...
paths:
  /user:
    get:
      operationId: list-User
      description: List "User" resources
      tags:
        - user
      parameters:
        - $ref: "#/components/parameters/include"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/page-offset"
        - $ref: "#/components/parameters/page-limit"
        - name: filter[id]
          required: false
          description: Id filter
          in: query
          style: form
          explode: false
          schema:
            type: integer
        - name: filter[email]
          required: false
          description: Equality filter for "email"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[email$contains]
          required: false
          description: String contains filter for "email"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[email$icontains]
          required: false
          description: String case-insensitive contains filter for "email"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[email$search]
          required: false
          description: String full-text search filter for "email"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[email$startsWith]
          required: false
          description: String startsWith filter for "email"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[email$endsWith]
          required: false
          description: String endsWith filter for "email"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[name]
          required: false
          description: Equality filter for "name"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[name$contains]
          required: false
          description: String contains filter for "name"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[name$icontains]
          required: false
          description: String case-insensitive contains filter for "name"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[name$search]
          required: false
          description: String full-text search filter for "name"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[name$startsWith]
          required: false
          description: String startsWith filter for "name"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[name$endsWith]
          required: false
          description: String endsWith filter for "name"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[role]
          required: false
          description: Equality filter for "role"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[role$contains]
          required: false
          description: String contains filter for "role"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[role$icontains]
          required: false
          description: String case-insensitive contains filter for "role"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[role$search]
          required: false
          description: String full-text search filter for "role"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[role$startsWith]
          required: false
          description: String startsWith filter for "role"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[role$endsWith]
          required: false
          description: String endsWith filter for "role"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[posts]
          required: false
          description: Equality filter for "posts"
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
      security: []
    post:
      operationId: create-User
      description: Create a "User" resource
      tags:
        - user
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
...
  /post:
    get:
      operationId: list-Post
      description: List "Post" resources
      tags:
        - post
      parameters:
        - $ref: "#/components/parameters/include"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/page-offset"
        - $ref: "#/components/parameters/page-limit"
        - name: filter[id]
          required: false
          description: Id filter
          in: query
          style: form
          explode: false
          schema:
            type: integer
        - name: filter[createdAt]
          required: false
          description: Equality filter for "createdAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[createdAt$lt]
          required: false
          description: Less-than filter for "createdAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[createdAt$lte]
          required: false
          description: Less-than or equal filter for "createdAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[createdAt$gt]
          required: false
          description: Greater-than filter for "createdAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[createdAt$gte]
          required: false
          description: Greater-than or equal filter for "createdAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[updatedAt]
          required: false
          description: Equality filter for "updatedAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[updatedAt$lt]
          required: false
          description: Less-than filter for "updatedAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[updatedAt$lte]
          required: false
          description: Less-than or equal filter for "updatedAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[updatedAt$gt]
          required: false
          description: Greater-than filter for "updatedAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[updatedAt$gte]
          required: false
          description: Greater-than or equal filter for "updatedAt"
          in: query
          style: form
          explode: false
          schema:
            type: string
            format: date-time
        - name: filter[title]
          required: false
          description: Equality filter for "title"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[title$contains]
          required: false
          description: String contains filter for "title"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[title$icontains]
          required: false
          description: String case-insensitive contains filter for "title"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[title$search]
          required: false
          description: String full-text search filter for "title"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[title$startsWith]
          required: false
          description: String startsWith filter for "title"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[title$endsWith]
          required: false
          description: String endsWith filter for "title"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[content]
          required: false
          description: Equality filter for "content"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[content$contains]
          required: false
          description: String contains filter for "content"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[content$icontains]
          required: false
          description: String case-insensitive contains filter for "content"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[content$search]
          required: false
          description: String full-text search filter for "content"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[content$startsWith]
          required: false
          description: String startsWith filter for "content"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[content$endsWith]
          required: false
          description: String endsWith filter for "content"
          in: query
          style: form
          explode: false
          schema:
            type: string
        - name: filter[published]
          required: false
          description: Equality filter for "published"
          in: query
          style: form
          explode: false
          schema:
            type: boolean
        - name: filter[author]
          required: false
          description: Equality filter for "author"
          in: query
          style: form
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
    post:
      operationId: create-Post
      description: Create a "Post" resource
      tags:
        - post
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/PostCreateRequest"
      responses:
        "201":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
  /post/{id}:
    get:
      operationId: fetch-Post
      description: Fetch a "Post" resource
      tags:
        - post
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/include"
      responses:
        "200":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
        "404":
          description: Resource is not found
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
    put:
      operationId: update-Post-put
      description: Update a "Post" resource
      tags:
        - post
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/PostUpdateRequest"
      responses:
        "200":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
        "404":
          description: Resource is not found
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
    patch:
      operationId: update-Post-patch
      description: Update a "Post" resource
      tags:
        - post
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/PostUpdateRequest"
      responses:
        "200":
          description: Successful operation
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
        "404":
          description: Resource is not found
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
    delete:
      operationId: delete-Post
      description: Delete a "Post" resource
      tags:
        - post
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Successful operation
        "403":
          description: Request is forbidden
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
        "404":
          description: Resource is not found
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/_errorResponse"
...
components:
  schemas:
    _jsonapi:
      type: object
      description: An object describing the server’s implementation
      required:
        - version
      properties:
        version:
          type: string
    _meta:
      type: object
      description: Meta information about the request or response
      properties:
        serialization:
          description: Superjson serialization metadata
      additionalProperties: true
    _resourceIdentifier:
      type: object
      description: Identifier for a resource
      required:
        - type
        - id
      properties:
        type:
          type: string
          description: Resource type
        id:
          type: string
          description: Resource id
    _resource:
      allOf:
        - $ref: "#/components/schemas/_resourceIdentifier"
        - type: object
          description: A resource with attributes and relationships
          properties:
            attributes:
              type: object
              description: Resource attributes
            relationships:
              type: object
              description: Resource relationships
    _links:
      type: object
      required:
        - self
      description: Links related to the resource
      properties:
        self:
          type: string
          description: Link for refetching the curent results
    _pagination:
      type: object
      description: Pagination information
      required:
        - first
        - last
        - prev
        - next
      properties:
        first:
          type: string
          description: Link to the first page
          nullable: true
        last:
          type: string
          description: Link to the last page
          nullable: true
        prev:
          type: string
          description: Link to the previous page
          nullable: true
        next:
          type: string
          description: Link to the next page
          nullable: true
    _errors:
      type: array
      description: An array of error objects
      items:
        type: object
        required:
          - status
          - code
        properties:
          status:
            type: string
            description: HTTP status
          code:
            type: string
            description: Error code
          prismaCode:
            type: string
            description: Prisma error code if the error is thrown by Prisma
          title:
            type: string
            description: Error title
          detail:
            type: string
            description: Error detail
    _errorResponse:
      type: object
      required:
        - errors
      description: An error response
      properties:
        jsonapi:
          $ref: "#/components/schemas/_jsonapi"
        errors:
          $ref: "#/components/schemas/_errors"
    _relationLinks:
      type: object
      required:
        - self
        - related
      description: Links related to a relationship
      properties:
        self:
          type: string
          description: Link for fetching this relationship
        related:
          type: string
          description: Link for fetching the resource represented by this relationship
    _toOneRelationship:
      type: object
      description: A to-one relationship
      properties:
        data:
          $ref: "#/components/schemas/_resourceIdentifier"
          nullable: true
    _toOneRelationshipWithLinks:
      type: object
      required:
        - links
        - data
      description: A to-one relationship with links
      properties:
        links:
          $ref: "#/components/schemas/_relationLinks"
        data:
          $ref: "#/components/schemas/_resourceIdentifier"
          nullable: true
    _toManyRelationship:
      type: object
      required:
        - data
      description: A to-many relationship
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/_resourceIdentifier"
    _toManyRelationshipWithLinks:
      type: object
      required:
        - links
        - data
      description: A to-many relationship with links
      properties:
        links:
          $ref: "#/components/schemas/_pagedRelationLinks"
        data:
          type: array
          items:
            $ref: "#/components/schemas/_resourceIdentifier"
    _pagedRelationLinks:
      description: Relationship links with pagination information
      allOf:
        - $ref: "#/components/schemas/_pagination"
        - $ref: "#/components/schemas/_relationLinks"
...
    User:
      type: object
      description: The "User" model
      required:
        - id
        - type
        - attributes
      properties:
        type:
          type: string
        id:
          type: string
        attributes:
          type: object
          properties:
            email:
              type: string
            name:
              type: string
            role:
              type: string
        relationships:
          type: object
          properties:
            posts:
              $ref: "#/components/schemas/_toManyRelationshipWithLinks"
    UserCreateRequest:
      type: object
      description: Input for creating a "User"
      required:
        - data
      properties:
        data:
          type: object
          description: The "User" model
          required:
            - id
            - type
            - attributes
          properties:
            type:
              type: string
            id:
              type: string
            attributes:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
            relationships:
              type: object
              properties:
                posts:
                  $ref: "#/components/schemas/_toManyRelationship"
        meta:
          $ref: "#/components/schemas/_meta"
...
    UserResponse:
      type: object
      description: Response for a "User"
      required:
        - data
      properties:
        jsonapi:
          $ref: "#/components/schemas/_jsonapi"
        data:
          allOf:
            - $ref: "#/components/schemas/User"
            - type: object
              properties:
                relationships:
                  type: object
                  properties: &a1
                    posts:
                      $ref: "#/components/schemas/_toManyRelationship"
        meta:
          $ref: "#/components/schemas/_meta"
        included:
          type: array
          items:
            $ref: "#/components/schemas/_resource"
        links:
          $ref: "#/components/schemas/_links"
    UserListResponse:
      type: object
      description: Response for a list of "User"
      required:
        - data
        - links
      properties:
        jsonapi:
          $ref: "#/components/schemas/_jsonapi"
        data:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/User"
              - type: object
                properties:
                  relationships:
                    type: object
                    properties: *a1
        meta:
          $ref: "#/components/schemas/_meta"
        included:
          type: array
          items:
            $ref: "#/components/schemas/_resource"
        links:
          allOf:
            - $ref: "#/components/schemas/_links"
            - $ref: "#/components/schemas/_pagination"
    Post:
      type: object
      description: The "Post" model
      required:
        - id
        - type
        - attributes
      properties:
        type:
          type: string
        id:
          type: string
        attributes:
          type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            title:
              type: string
            content:
              type: string
            published:
              type: boolean
            authorId:
              type: integer
        relationships:
          type: object
          properties:
            author:
              $ref: "#/components/schemas/_toOneRelationshipWithLinks"
    PostCreateRequest:
      type: object
      description: Input for creating a "Post"
      required:
        - data
      properties:
        data:
          type: object
          description: The "Post" model
          required:
            - id
            - type
            - attributes
          properties:
            type:
              type: string
            id:
              type: string
            attributes:
              type: object
              required:
                - updatedAt
                - title
              properties:
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time
                title:
                  type: string
                content:
                  type: string
                published:
                  type: boolean
                authorId:
                  type: integer
            relationships:
              type: object
              properties:
                author:
                  $ref: "#/components/schemas/_toOneRelationship"
        meta:
          $ref: "#/components/schemas/_meta"
    PostUpdateRequest:
      type: object
      description: Input for updating a "Post"
      required:
        - data
      properties:
        data:
          type: object
          description: The "Post" model
          required:
            - id
            - type
            - attributes
          properties:
            type:
              type: string
            id:
              type: string
            attributes:
              type: object
              properties:
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time
                title:
                  type: string
                content:
                  type: string
                published:
                  type: boolean
                authorId:
                  type: integer
            relationships:
              type: object
              properties:
                author:
                  $ref: "#/components/schemas/_toOneRelationship"
        meta:
          $ref: "#/components/schemas/_meta"
    PostResponse:
      type: object
      description: Response for a "Post"
      required:
        - data
      properties:
        jsonapi:
          $ref: "#/components/schemas/_jsonapi"
        data:
          allOf:
            - $ref: "#/components/schemas/Post"
            - type: object
              properties:
                relationships:
                  type: object
                  properties: &a2
                    author:
                      $ref: "#/components/schemas/_toOneRelationship"
        meta:
          $ref: "#/components/schemas/_meta"
        included:
          type: array
          items:
            $ref: "#/components/schemas/_resource"
        links:
          $ref: "#/components/schemas/_links"
    PostListResponse:
      type: object
      description: Response for a list of "Post"
      required:
        - data
        - links
      properties:
        jsonapi:
          $ref: "#/components/schemas/_jsonapi"
        data:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/Post"
              - type: object
                properties:
                  relationships:
                    type: object
                    properties: *a2
        meta:
          $ref: "#/components/schemas/_meta"
        included:
          type: array
          items:
            $ref: "#/components/schemas/_resource"
        links:
          allOf:
            - $ref: "#/components/schemas/_links"
            - $ref: "#/components/schemas/_pagination"
  parameters:
    id:
      name: id
      in: path
      description: The resource id
      required: true
      schema:
        type: string
    include:
      name: include
      in: query
      description: Relationships to include
      required: false
      style: form
      schema:
        type: string
    sort:
      name: sort
      in: query
      description: Fields to sort by
      required: false
      style: form
      schema:
        type: string
    page-offset:
      name: page[offset]
      in: query
      description: Offset for pagination
      required: false
      style: form
      schema:
        type: integer
    page-limit:
      name: page[limit]
      in: query
      description: Limit for pagination
      required: false
      style: form
      schema:
        type: integer